Basic layout for page is done.
