+++ src/data/jobListings.json
@@ -6,7 +6,7 @@
     "address": "343 W Erie #310 Chicago, IL 60654",
     "featured": true,
     "position": "Platform Engineer",
-    "company": "Drafbit",
+    "company": "Draftbit",
     "datePosted": "2018-12-03T15:18:31.433Z",
     "neighborhood": "River North",
     "technologies": ["Javascript", "Node.js", "React.js", "React Native", "Ruby", "GraphQL"],
@@ -20,7 +20,7 @@
     "address": "343 W Erie #310 Chicago, IL 60654",
     "featured": true,
     "position": "Platform Engineer",
-    "company": "Drafbit",
+    "company": "Draftbit",
     "datePosted": "2018-12-03T15:18:31.433Z",
     "neighborhood": "River North",
     "technologies": ["Javascript", "Node.js", "React.js", "Gatsby"],
