+++ src/pages/job-listings/index.js
@@ -2,27 +2,8 @@ import React from 'react'
 import { graphql } from 'gatsby'
 import { Link } from 'gatsby'
 import Layout from '../../components/Layout'
+import { renderTechIconCorrectUrl } from '../../utils/index'
 import './job-listings.css'
-
-const renderCorrectUrl = tech => {
-  let lowerCase = tech.toLowerCase()
-  let parsedTech = ''
-  switch (lowerCase) {
-    case 'react.js':
-      parsedTech = 'react'
-      break
-    case 'react native':
-      parsedTech = 'react'
-      break
-    case 'node.js':
-      parsedTech = 'node-dot-js'
-      break
-    default:
-      parsedTech = lowerCase
-      break
-  }
-  return `https://cdn.jsdelivr.net/npm/simple-icons@latest/icons/${parsedTech}.svg`
-}
 
 const JobPost = ({
   postID,
@@ -65,7 +46,7 @@ const JobPost = ({
               width="20"
               height="20"
               className="rounded mr-3"
-              src={renderCorrectUrl(tech)}
+              src={renderTechIconCorrectUrl(tech)}
               alt={`Logo for ${tech}`}
             />
           ))}
